requirements.txt
torch
pandas
numpy
matplotlib
seaborn
scikit-learn
Pillow
torchvision
flask

